NP-18140 Show and hide registration type selector
